requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pymongo==4.7.2
python-dotenv==1.0.1
